[Java] Add array analysis to ATJ.
This is realized as a pre-translation step (even though it currently does not
modify the ACL2 code being (pre-)translated), between the type annotation step
and the variable reuse step -- see the developer documentation for why that's
the right or best place to perform this array analysis.

This is a stobj-style single-threadedness analysis for (ACL2 representations of)
Java arrays. The commonality and differences between this analysis and the stobj
analysis performed by ACL2 are discussed in the developer documentation of the
array analysis code.

This array analysis is designed so that it should ensure the safety of
destructively updating arrays in the generated Java code (an upcoming extensions
of ATJ's translation step). This remains to be formally proved.