[Java] Extend ATJ to generate destructive Java array updates.
Now ATJ recognizes the ACL2 functions that model array write operations and
translates them to Java calls of methods that destructively update arrays. The
correctness of this is justified by the array analysis added in the previous
commit (although that remains to be proved formally).

We translate the ACL2 array write function calls to calls of methods that
destructively update the arrays and then return the updated arrays for now, so
that we maintain the "uniformity" of ATJ's ACL2-to-Java translation step, in
which every ACL2 term and function is turned into a Java expression and method
that returns the Java counterpart of the ACL2 result. But in the Java code,
returning the arrays is unnecessary, since they get destructively
updated. However, having certain ACL2 terms and functions be translated to Java
expressions and methods that do not return a result (i.e. expressions turned
into statements without using their value, and void methods) introduces a
complication in the ACL2-to-Java translation that may be best avoided there,
adding instead a Java-to-Java post-translation step that removed the returning
of the updated arrays. This will be investigated next.