Tweaked comments in check-system-guards.lisp.  Updated auto-termination data files.  Updated :doc note-8-3 anticipating release soon.  Fixed untranslated-theoremp in support of pseudo-good-worldp.  Changed a keyword argument in definductor from :rel to :well-founded-relation.
Tweaked comments in check-system-guards.lisp (and ran the longer check
suggested in a comment there).

Updated kestrel/auto-termination/td-cands.{acl2,lisp} as per the
README there.

Updated :doc note-8-3 with release date of this month and with a
comment at the top of the defxdoc event about the number of release
note items of each type.

Fixed an incorrect earlier change to untranslated-theoremp in
books/system/pseudo-good-worldp.lisp.

(As requested by J) Modified definductor so that its keyword argument
is :well-founded-relation rather than :rel, to match what's used in
xargs.