[Java] Extend Java language formalization.
Add a model of 'Unicode input characters'. These, following the grammar, are
either raw (i.e. "plain") Unicode characters, or Unicode escapes. We introduce
this explicit concept with the purpose of using it in some parts of the abstract
syntax, so that we can have a more accurate representation of a Java program,
particularly for code generation purposes. Distinguishing raw vs. escape makes
sense in Java identifiers, in Java character literals, and in Java string
literals.