[APT] Extend ISODATA.
Generate the theorem that rewrites (calls of) the new function to (terms
involving calls of) the old function non-locally, like the theorem that rewrites
the old function to the new function. For ISODATA, these two theorems have a
symmetric structure.

A new input, :NEW-TO-OLD, has been added to ISODATA to specify the name to
use. This introduces a new, more general mechanism to specify this kind of
theorems names, compared to the paired name utilities used before. The mechanism
involves the APT defaults table.

The :THM-NAME input will be soon renamed to :OLD-TO-NEW, for symmetry. There
will be additional changes to make things uniform with ISODATA.