[Java] Extend the ATJ tutorial.
Add a page on the shallow embedding approach.

This provides a general overview of the approach, and describes the kind of Java
code that is generated "by default". Several upcoming tutorial pages will
describe the rich features supported by the shallow embedding, such as the use
of types, the targeting of Java primitive values and arrays, the destructive
updating of arrays, etc.