[Java] Add an ATJ post-translation step.
This new transformation step, which take place within Java, replaces expression
statements of the form

  a = writeArray(a, i, x);

with statements of the form

  a[i] = x;

The former arise from the translation of ACL2 terms of the form

  (let (... (a (write-array a i x)) ...) ...)

where the call of the write-array function is translated to the call of the
writeArray method, for uniformity of translation. (There are actually 8
write-array functions and 8 writeArray methods, one and one for each of Java's 8
primitive types.)

The writeArray method internally executes the assignment a[i] = x and then
returns a, which is correct because of the single-threadednes array analysis.
So this post-translation steps essentially "inlines" the method call, but it
actually replaces the whole assignment, not just the call. This is correct
because the array a is the same on the left-hand side of the assignment and as
the first argument of the writeArray method, which is guaranteed by the array
analysis.

This eliiminates one kind of writeArray calls. Future post-translation steps
will eliminate other kinds.