[APT] Improve ISODATA.
Add an :OLD-TO-NEW input that will eventually replace the :THM-NAME input.
Temporarily, these two inputs coexist, with the latter taking precedence over
the former. As soon as all the files that use ISODATA (some of which are outside
the community books) no longer use :THM-NAME, this input will be removed.

Note that :OLD-TO-NEW is consistent with the :NEW-TO-OLD input. They refer to
two "symmetric" theorems generated by ISODATA.