[FTY] Change the package of some files.
These are the files for the individual standard instances of DEFBYTE. It looks
like if these files are in the FTY package then the manual pages for the
fixtypes, recognizers, and fixers (which are all in the ACL2 pacakge), get an
'ACL2::' qualifier in front. Putting the files in the ACL2 package should
eliminate the qualifier.