[Java] Extend an ATJ post-translation step.
This is the Java-to-Java post-translation step that removes certain calls of the
array write methods. Now we remove additional ones. Every statement of the form

  return <array-write-method>(a, i, x);

is replaced with two statements

  a[i] = x;
  return a;

Some of the existing tests in [books]/kestrel/java/atj/tests/primarrays.lisp
demonstrate this new capability in the generated Java code.