[Java] Improve ATJ's array analysis.
Require arguments of named functions that are arrays to be variables. That is,
we disallow "nested" calls of array-modifying functions. Instead, the ACL2 code
to be translated to Java must assign each call to the appropriate array
variable, similarly to ACL2's restrictions on stobjs.

This change is motivated not by correctness, but to facilitate the subsequent
post-translation step that "inlines" the array-modifying functions to make the
Java code more idiomatic.

With this change, now all the calls of the eight "basic" array writing methods
should be inlined during post-translation.