[APT] Improve ISODATA.
Replace the :THM-ENABLE input with a new :OLD-TO-NEW-ENABLE input,
which is more consistent with the existing inputs and with an upcoming
input. This new :OLD-TO-NEW-ENABLE input, if not explicitly supplied,
is taken from the APT defaults table.

Adjust the tests and the workshop examples, given that the initial
default value of the new input is NIL, i.e. disable the theorem.

Also reduce the book inclusions of the workshop loop example.

Also use MUST-BE-REDUNDANT in the workshop integer example.