[Java] Improve ATJ's model of Java primitive arrays.
Tweak array write functions so that they unconditionally preserve the
number of components of the arrays.

Add linear rules about the upper bound of the number of components of
an array.

Add rules showing that array writes preserve the predicate that says
whether a Java int is a valid index for an array. These may be useful
in guard proofs involving array operations.