[Java] Improve ATJ's handling of arrays.
Extend ATJ's model of the Java primitive arrays with new functions to convert
between arrays and corresponding ACL2 lists. The code generator recognizes these
functions and translates them to the appropriate Java conversion code.

The Java primitive array deconstructors introduced by the fixtype are no longer
recognized and translated to conversion code: that was in fact a bug, since the
returned list would have to be one of Java primitive values, not of their ACL2
counterparts.

Extend ATJ's model of the Java primitive arrays with new functions to create
arrays from components. These are essentially synonyms of the fixtype
constructors, but provide more flexibility in case we want to change the
definitions of the fixtypes slightly. The code generator only accepts calls of
these functions on arguments that are LIST calls. It no longer generates
conversion code when the argument is not a LIST call: that was in fact a bug,
for the same reason mentioned above for the deconstructors.