Tweaked discussion of :restrict in :doc hints.  Then synched doc.lisp etc.
This change incorporates a new example sent to me by Mihir Mehta,
replacing (as per his suggestion) the example he added earlier today.
It's a simpler example than the other, so I switched the order,
putting the new one before the existing one (the one that was from
Bishop Brock), which I think is still informative.  While I was at it,
I made some other tweaks to the discussion of :restrict.