Improved Cryptic BRR Message 2 (and also 1) to clarify that the behavior may occur from a recursive prover call.  Improved an error message in the case of the proof-builder EQUIV or = command, when a macro alias is supplied for an equivalence relation.  Fixed :doc verify-guards-for-system-functions to include certification of system/apply/loop-scions.cert.
Quoting comments in (defxdoc note-8-4 ...):

; Any ``Cryptic BRR Message'' printed by the prover now acknowledges that the
; issue may be due to an invocation of the prover from within BRR.  (We do not
; recall BRR being designed to behave well in this case.)  :DOC brr has been
; updated accordingly.  Thanks to Mihir Mehta for sending an example that led
; to this improvement.

; Improved the error messages from the proof-builder EQUIV and = commands in
; the case that the alleged equivalence relation is actually a macro alias for
; an equivalence relation.  Thanks to Mihir Mehta for sending an example that
; led to this improvement.  Also avoided a raw Lisp error for EQUIV when the
; alleged equivalence relation is supplied as an atom that is not a symbol.

Thanks to J for helping me to work through the issue described in the
first comment above.  The new message makes it clear that there is no
reason to send the implementors an example when the message is printed
after making a call of THM etc. from within a BRR break.  Here is the
new "Cryptic BRR Message 2".  It is the same as the new "Cryptic BRR
Message 1" except for the first sentence.

Cryptic BRR Message 2:  Discarding dead brr-stack.  If you are seeing
this message after you have caused a console interrupt (or otherwise
have made a break into Lisp) or during a new prover call made from
within BRR, then it is likely that this is a bit of routine BRR housekeeping.
(In the latter case, of a new prover call made from within BRR, you
can probably avoid this message by calling :brr nil before making that
new prover call.)  Otherwise, the ACL2 implementors would appreciate
a script showing how to reproduce this message.

For the issue described in the first comment above, the error message
now mentions the equivalence relation, e.g.:

->: (= (NFIX M) M :EQUIV MY-EQUIV)
*** NO CHANGE *** -- The ``equivalence relation'' that you supplied,
MY-EQUIV, is not known to ACL2 as an equivalence relation.  Perhaps
you intended the corresponding function for which it is an ``alias''(see
:DOC macro-aliases-table), MY-EQUIV$INLINE.
->:

Finally, regarding fixing of :doc verify-guards-for-system-functions
to include certification of system/apply/loop-scions.cert: the
corresponding instructions in a comment in
*system-verify-guards-alist* (source file boot-strap-pass-2-b.lisp)
had already been updated some time ago to include that, but Mihir
Mehta pointed out that this update was missing from :doc
verify-guards-for-system-functions (thanks, Mihir).