Helper functions position-ac-eq-exec, position-ac-eql-exec and position-equal now fix their accumulator argument.
Quoting :doc note-8-4:

  Functions position-ac-eq-exec, position-ac-eql-exec, and
  position-equal-ac, which all support the macro, [position], now fix
  their accumulator argument.  Thanks to Mihir Mehta for supplying
  these changes, for the purpose of avoiding [ACL2-numberp] type
  hypotheses.

Thanks to Mihir for running an "everything" regression as well as
"make proofs" and "make devel-check" when submitting this change.  The
change to books/misc/beta-reduce.lisp was his as well, made to reflect
the changes mentioned above.

[Note: Normally we might follow the process in :doc
developers-guide-contributing.  But Mihir and I agreed that in this
case, it was good for him simply to send me the changes, for me to
commit and push.  I ran an "everything" regression myself before
committing.]