Temporarily add file for community books talk.
As done for the previous workshop (ACL2-2018), I'm temporarily adding this file
here to facilitate additions and edits by the community book contributors. I'll
take care of the overall organization of these slides, and present them at the
workshop, as asked by the organizers. This file will be removed after the
workshop (as always, there will be a PDF posted on the workshop web site).