[Java] Improve ATJ tutorial.
Avoid the explicit distinction between 'main' and 'auxiliary' pages. Instead,
talk about a "thread" of Next/Previous links through a subset of the pages, to
be followed when reading the tutorial for the first time. The pages not in that
threaded subset are reachable through other links from the threaded pages.

All the pages are alphabetically ordered subtopics now, for easier reference.