Added documentation pertaining to useless-runes.  Fixed bug in ACL2(p) handling of useless-runes.  Made minor comment edits in two cert.acl2 files.
My preceding commit message noted that documentation for useless-runes
was yet added.  That is fixed now:

- documented :useless-runes option in :doc certify-book;
- added :doc useless-runes;
- improved :doc accumulated-persistence;
- noted useless-runes limitation in :doc
    unsupported-waterfall-parallelism-features when
    waterfall-parallelism is enabled; and
- updated :doc note-8-4.

Fixed a bug that prevented writing of @useless-runes.lsp files by
ACL2(p) with waterfall-parallelism disabled.  Added to :doc
unsupported-waterfall-parallelism-features to explain that reading and
writing of the @useless-runes.lsp file when waterfall parallelism is
enabled.