[Java] Include ATJ types for built-in functions automatically.
This reduces the chances of users forgetting to include these. Now they come
with ATJ.

Forgetting to include those does not cause any error, but results in worse Java
code in general.

Even though they are only relevant to the shallow embedding with guards, this is
the more typical mode in which ATJ should be used.