[Java] Add secondary ATJ types for the built-in NOT.
NOT's primary ATJ input and output types are :AVALUE (input) and :ABOOLEAN
(output). This results in the generation (in the shallow embedding with guards)
of a Java method not(Acl2Value) that returns boolean.

By adding secondary ATJ input and output types :ABOOLEAN (input) and :ABOOLEAN
(output) for NOT, an overloaded not(boolean) method is now generated (also
returning boolean). This improves the quality of the code, because when NOT is
applied to a boolean in ACL2, the corresponding Java code will select
not(boolean) and thus avoid converting the Java boolean to the Acl2Symbol for T
or NIL, which is correct but not as readable and efficient.

This issue arises because of a recent improvement to ATJ where ACL2 booleans get
mapped to Java booleans.

Thanks to Limei Gilham for an exmaple that showed this issue.