Commented out a couple of tests that fail in SBCL 2.0.5.
Here is the comment I left above the two tests in
books/std/io/read-string-tests.lisp that I excluded for SBCL.

;; Update from Matt K., 5/31/2020: SBCL 2.0.5 has the problem discussed above,
;; so I am disabling these tests once again for SBCL.  I think we are actually
;; hitting a bug in function read-string-fn in read-string-raw.lsp (stream has
;; dynamic extent); probably I'll fix it tomorrow and then restore SBCL here
;; (and delete this comment).

To elaborate a bit: The problem (thanks to Stas Boukarev for pointing
this out, in a reply to my message to sbcl-bugs) is that string
streams have dynamic extent (indeed, this is documented in the
HyperSpec), so the error message printed by read-string-fn (defined in
books/std/io/read-string-raw.lsp) fails because the relevant string
stream, referenced (implicitly) in that message, has already been
freed.  In other Lisps supporting ACL2, and that stream is not
referenced in that error message.

One other thing: As you can guess from the comment above, for this
regression I used the new release SBCL 2.0.5 that came out today.  As
a result, there was a hard Lisp error from the event
resolve-b+-sweep-into-pp+ in
books/projects/rp-rewriter/lib/mult/meta/sum-meta.lisp.  To me it
looks like it may be a compiler bug, but we'll see; I've sent a
message to the sbcl-help list (with a much simplified example of the
problematic behavior).  For now I am leaving that book unchanged.  A
workaround, if necessary, might be to define function
does-belong-to-pp+ using defund instead of defund-inline.  I'll keep
on top of this.