[Java] Improve the ATJ tutorial.
Now all the pages are linked in a sequence of next/previous links that provide a
sequence in which to read the tutorial the first time. The distinction between
main and auxiliary pages, which was somewhat artificial, is gone; but some pages
start with text saying that they may skipped at first reading.

Furthermore, macros have been added to define the content of the pages and to
generate XDOC topics from them. This way, the pages can be provided in the
desired order for the next/previous links, which are automatically generated.