[Java] Improve code generated by ATJ.
Now, in the shallow embedding with guards, ACL2 characters are translated to
Java char values when possible, and automatically converted to/from
Acl2Character objects when needed.

This is analogous to the mapping of ACL2 booleans to Java booleans, recently
implemented.

Even though ACL2 characters (which consist of 8 bits) are not isomorphic to Java
characters (which consist of 16 bits), under the assumptions of guard
satisfaction (i.e. when the :GUARDS input of ATJ is T), and provided that
external Java code calls (Java methods generated from) ACL2 functions with
characters below 256, the Java characters manipulated by the generated Java code
should always be below 256.