[Java] Extend AIJ.
Add variant Java implementations of some natively implemented ACL2 functions
that take and/or return Java strings instead of ACL2 strings.

This is in support of an upcoming extension to ATJ that, in the shallow
embedding with guards, will map ACL2 strings to unwrapped Java strings when
possible, analogously to what was done for booleans and characters previously.