[APT] Add EXPDATA, the expanded data transformation.
This transformation is useful, like ISODATA, for data type refinements. But it
supports richer mappings than isomorphic ones: each old datum may be represented
by multiple new data. Instead of requiring an isomorphism between old and new
data, we require a surjection from the new data to the old data. This is a data
abstraction function, in typical data type refinement nomenclature.

We also require the existence of a right inverse for the surjection, which is
used as the forward conversion. In fact, currently the data transformation is
essentially the same as the one performed by ISODATA, except that the
conversions are not necessarily isomorphic. It is likely that we will need to
extend EXPDATA to allow the use of different right inverses in different places
of the new function (given that the surjection is not necessarily injective,
there are in general multiple right inverse). We already have a preliminary
design for these extensions.

The current version of EXPDATA includes user documentation, design notes, and
tests.