[Java] Improve code generated by ATJ.
In the shallow embedding with guards, now calls of ACL2's AND are translated to
Java's && expressions when possible, resulting in more idiomatic and efficient
Java code. This happens automatically. This is separate from the explicit
targeting of Java's primitive operators (such as &&) via refinement to an ACL2
form that uses the ACL2 models of the Java operators.