[Java] Improve code generated by ATJ.
Now the code to build the ACL2 environment (i.e. the ACL2 package definitions,
and in the deep embedding approach also the ACL2 function definitions) is
generated in a separate class in a separate file. This code is fairly large
(roughly 50k lines), but it is boilerplate and "uninteresting". This way, the
main Java class and file now contain the "interesting" code. This also avoids
some decreased responsiveness of Emacs when moving around the previous large
monolithic file.

Update user documentation accordingly.

Update tutorial accordingly.

Update .java ignored files to add the newly generated files for the tests.