[Java] Improve code generated by ATJ.
Now, in the shallow embedding approach, the main class starts off with the
static initializer and initialize() method, followed by the nested classes
corresponding to the ACL2 packages. The nested classes are followed by the final
static fields for the quoted constants, then by the nested classes for the ACL2
multiple values, and then by the array conversion methods. This way, the
"interesting" code is more readily visible when opening the file.