[Java] Improve the code generated by ATJ.
Now, in the shallow embedding, we generate wrapper methods not for all the
natively implemented ACL2 functions, but only for the ones actually used,
i.e. directly or indirectly called by the target functions.

This involved some slight changes to the worklist algorithm that computes the
closure of the functions to translate to Java, and some adaptation to some code
generation functions.