[Java] Improve the code generated by ATJ.
In the shallow embedding approach, now the generated methods declare to throw
Acl2UndefinedPackageException only if they can actually throw it (before every
method was declared to do so, for simplicity). Only the methods that call,
directly or indirectly, the native Java implementations of PKG-WITNESS and
PKG-IMPORTS can throw this exception.

This has been accomplished by extending the worklist algorithm (which calculates
the call closure of the target functions) to also return a call graph. Then the
direct and indirect callers of PKG-WITNESS and PKG-IMPORTS are obtained from the
graph.

In the future, this can be extended to a mode general fixpoint analysis on the
graph to determine which exceptions can be thrown by which functions. Once we
extend ATJ to accept ACL2 functions with side effects, which involves suitably
replicating those side effects in Java, we may have a richer range of exceptions
thrown by the Java methods that the ACL2 functions translate to (e.g. I/O
exceptions).