[ACL2 PL] Extend formalization a bit.
Add a function that returns the free variables of a (translated) term. Note
that, in this formalization, for greater flexibility we do not require lambda
expressions to be closed.

Also add a fixtype of osets of symbol values. The free variables of a term are
returned as a value of this fixtype.