CMUCL builds work again: we avoid destructive modification of constants.  Clarified in installation instructions that "make basic" certifies only some books.  Fixed defstobj bug for Lisps other than CCL and SBCL.  Sped up certification time for one book in (at least) Allegro CL.
Quoting :doc note-8-4:

  ACL2 can once again be built on CMU Common Lisp (CMUCL) (though we
  have only done minimal testing).  The problem turned out to be with
  ACL2, not CMUCL: low-level Lisp code in the ACL2 sources was
  destructively modifying a quoted constant.  (For implementation
  details, see *fncall-cache* in source file translate.lisp.)  The
  bug was discovered when considering modification of the build
  process to compile ACL2 source files when the host Lisp is SBCL.
  Thanks to Stas Boukarev for pointing us in the right direction to
  debug this error.

Quoting a related comment in (defxdoc note-8-4 ...):

; We changed Lisp variable *inside-absstobj-update* so that it is no longer
; initialized with a constant.  This seemed potentially important given the
; item below about destructively modifying a quoted constant in *fncall-cache*.

Thanks to David Russinoff for feedback about the installation
instructions.

My recent introduction of defstobj keyword :non-executable resulted in
a bug that could show up in Allegro CL and CMUCL (and probably
LispWorks and GCL), but maybe not CCL or SBCL, which don't compile
on-the-fly.  (Technical note: The bug was in the cltl-cmd argument to
install-event in defstobj-fn, which caused add-trip to fail to grab
the expected event-form.)  I fixed this bug.

Sped up certification of books/tools/oracle-time-tests.lisp by adding
(comp t).  I've seen it previously take over an hour for Allegro CL,
where now a dozen seconds suffices.  This shouldn't make any
difference for CCL or SBCL since they compile on-the-fly, but it could
be helpful for all other Lisps.

Successfullly ran "make regression-everything" in SBCL and, before an
update, successfully ran "make basic" in CCL, CMUCL, and Allegro CL.