[Java] Improve the code generated by ATJ.
In the shallow embedding, avoid generating synonym methods in classes for
packages that would otherwise have no methods (i.e. such that none of the
functions to translate to Java are in that package).

This greatly reduces the size of the generated Java code, given that there are,
unavoidably, fairly numerous ACL2 packages loaded in the ACL2 world at any time
in which ATJ is invoked.

ATJ still generates code to construct a representation of all the ACL2 packages
known at that point. But this is in the environment class, and happens in both
the deep and shallow embedding. The code reduction achieved by this commit
refers, instead, to the Java classes generated for the packages of the functions
being translated to Java methods; this applies to the shallow embedding only.