[Java] Improve the code generated by ATJ.
In the tail recursion elimination post-translation step, avoid generating
trivial assignments (in the body of the loop) of the form var = var. These
naturally arise when a formal parameter of the tail-recursive function is
passed, unchanged, as actual parameter to the recursive call. Thus we add a
check to exclude these assignments if they arise.