[APT] Improve the TAILREC transormation.
Now the automatically generated names for the new and wrapper function (when
:NEW-NAME is :AUTO and/or :WRAPPER-NAME is :AUTO) are as follows. If the wrapper
is not generated, the new name is obtained by adding or suitably incrementing
the explicit or implicit index of tha target function; this is as before. But if
the wrapper is generated, then that name is used for the wrapper, while the new
function has the form ...-AUX{...}, as common when a function (the wrapper in
this case) needs a "loop". The conjecture behind this change is that most of the
time we do not generate the wrapper, unless the function being transformed is
one of the top-level functions, in which case we want the wrapper, which has the
same interface as the target function, and we want it to have a name that
relates directly to the old function. This approach may be revisited in the
future. Also, in the future we may support more flexible naming schemes than the
form ...-AUX{...} for the generated tail-recursive function when there is a
wrapper.

Update the user documentation accordingly.

Update some tests accordingtly.