[APT] Improve the TAILREC transformation.
Change the default of the :WRAPPER input from T to NIL.

The rationale is that it is expected that the wrapper will be generated less
frequently. Unless the function to which TAILREC is applied is a "top-level"
function (e.g. an interface function of a program being transformed), just
generating the tail-recursive version, with the additional accumulator argument,
should be sufficient.

Update user documentation accordingly.

Update tests accordingly.