[SOFT] Add DEFINE-SK2.
This is DEFINE-SK followed by DEFSOFT. It is in a separate file to avoid
imposing the inclusion of DEFINE-SK. (On the other hand, DEFINE2 is not in a
different file because SOFT already includes DEFINE for its own implementation.)

DEFSOFT and DEFUN-INST treat DEFINE-SK2 the same as DEFUN-SK2 for now, but
future versions may do more by taking advantage of the additional
structure. Nonetheless, DEFINE-SK2 already provides more convience compared to
DEFUN-SK2.