Improved translation of function calls in the presence of congruence stobjs, in particular improving error messages.  Added comments and checks regarding *blacklisted-apply$-fns*.  Fixed :doc print-gv.  Fixed reference to a stub file.
Quoting :doc note-8-4:

  Improved translation of function calls, especially those that involve
  congruent [stobj]s, including better error messages, much improved
  code comments, and simplified code.  Thanks to Sol Swords for
  sending an example with a misleading error message.

The constant *blacklisted-apply$-fns* is a list of functions that
cannot be warranted.

Fixed two problems with :doc print-gv.  One was a markup typo.  The
other was a failure to point out that the output from print-gv always
goes to the terminal.

The stub file replacement was std/testing/eval for misc/eval, in
the book projects/apply/loop-recursion-examples.lisp.