Be more flexible in loading libipasirglucose4.so
The doc topic IPASIR::BUILDING-AN-IPASIR-SOLVER-LIBRARY claims that to
get the ipasir shared library working, it is sufficient to set
$IPASIR_SHARED_LIBRARY to the library's filename alone, and put its
directory in the $LD_LIBRARY_PATH list.

Unfortunately, while the loading code in load-ipasir-sharedlib-raw.lsp
itself is able to handle this, features.sh is stricter and requires
$IPASIR_SHARED_LIBRARY to contain a full path to the shared library,
so running `make regression` will skip the ipasir books if the library
is pointed to by a directory path in $LD_LIBRARY_PATH and a filename
in $IPASIR_SHARED_LIBRARY.

This commit generally expands the different ways in which the user is
able to provide an ipasir shared library for the ipasir-related books
to build from.  See the changes made to the aforementioned doc topic
for more details.