[Std/util] Improve DEFMAPPING.
Improve the robustness of the generated proofs by preventing the expansion of
MV-NTH in the empty theory, via a system attachment to the recently added hooks
to customize the expansion of MV-NTH in proofs when it is disabled.

This change, and in fact the new feature to customize the expansion of MV-NTH,
are motivated by an example that used to fail in DEFMAPPING, because of the
unwanted expanstion of MV-NTH in a generated proof.

Add, as a test, the aforementioned example, which now works.