[APT] Improve DIVCONQ.
Change the :SCHEMA input to have no default. Since this transformation will be
extended to support more schemas besides the current :LIST-FOLD schema, and no
schema is presumably more common than others, it makes sense to always be
explicit.

A separate thread is to also support an :AUTO schema that heuristically tries to
find the best schema for the problem at hand. This will be investigated
eventually.