[ecurve] Add fixtype support for points.
Currently the recognizer of (all possible) points of (all possible) elliptic
curves (over prime fields) is defined "directly", i.e. via a fixtype.

This commit adds a new file that provides a fixer and a fixtype for that
recognizer, as well as functions and theorems that should enable the treatment
of elliptic curve points as if they were introduced via FTY::DEFTAGSUM. I did
not attempt to replicate all the events that FTY::DEFTAGSUM would have
generated, but the main ones should be there. This may be extended as needed.

Using this fixtype support should generally ease guard and return type proofs in
ACL2 code that manipulates elliptic curve points.