[APT] Add a new SOLVE transformation.
This transformation attempts to solve a synthesis (i.e. constraint satisfaction)
problem directly, via inference techniques. The current implementation is just a
small start in that direction: it supports the Axe rewriter as inference method
(usable only at Kestrel until Axe is open-sourced). But it will support the ACL2
rewriter soon. It also suppors a "backup" manual method in which the user
supplied the solution directly and has the responsibility to prove it: this is
still more convenient than manually writing the whole refinement step by hand,
because also this manual solving method auto-generates several functions and
theorems with proofs.

The inference tool, the Axe rewriter in this case, is called via reflection, to
avoid having this transformation necessarily depend on Axe and other inference
tools. Only if the tools are included (via INCLUDE-BOOK), they are available for
use.

Not all synthesis problems can be solve this way, of course, but some are,
particularly the ones at the "leaves" of a top-down synthesis problem
decomposition.

There are plans to support the use of SMT and SAT solvers too. Thus, this is
moving towards techniques of synthesis by sketching, which can be a valuable
tool in the APT toolkit.

Thanks to Eric Smith for explaining how to use programmatically the Axe rewriter
and how to intepret its results.