[APT] Add some design notes and some documentation.
The design notes describe specification and refinements in general, and then the
pop-refinement and shallow pop-refinement approaches. They also describe various
forms of shallow pop-refinement specifications.

The new manual page documents the aforementioned forms of specification, for
reference in the user documentation of APT transformations that expect
specifications in those forms.