[FTY] Improve DEFSET.
Ensure that all the generated theorem names include the recognizer of the set
fixtype in their names, thus greatly increasing the chances of the theorem not
being already in use.

Before this improvements, it was not possible to create two set types with the
same element type, because DEFSET was generating certain theorem names that only
included the recognizer of the element type but not the name of recognizer of
the set type.

Thanks to Eric McCarthy for reporting this issue.