[C] Extend ATC to generate functional correctness proofs.
For every generated C function, we also generate a theorem asserting that, under
the guard of the corresponding ACL2 function, the execution of the C function
according to the formalized dynamic semantics returns the same result as the C
function.

Thus, ATC is now a proof-generating C code generator for ACL2. The current
subsets of ACL2 and C that it handles are still small and simple, but it is a
start.