[ATC] Support the representation of C tests.
In C, tests (e.g. of conditionals) are integer-valued expressions, where 0 means
false and non-0 means true. To represent C tests shallowly embedded in ACL2, we
introduce an ACL2 predicate that takes a C int value and return T or NIL based
on whether the C int value is non-0 or 0.