Clarified in :doc that proof-builder commands = and EQUIV don't use patterned congruences.
Thanks to Mihir Mehta for a Slack query that led to these :doc improvements.

It was already documented that the prover doesn't use patterned
congruences to do equality substitutions "at the goal level".  (It
notes that they are indeed used in rewriting.)  Maybe someday
patterned congruences will be used in the proof-builder = and EQUIV
commands and in more prover processes (specifically, in fertilization
and in removing trivial equivalences).  But at this point those seem
like low priorities.