Added comment near top of books/system/pseudo-good-worldp.lisp inviting people to contribute.
Thanks to Eric Smith and Alessandro Coglio for encouraging this
invitation (and in particular, to Eric for helping to get it right).

An improved pseudo-good-worldp predicate could help to simplify
statements of guards, for example, a guard specifying that every table
is an alist and that certain state globals are bound to appropriate
sorts of values: instead of conjoining such conditions,
(pseudo-good-worldp wrld t) would suffice.  A good example might be
include-book-dir.