Enhanced the prove-guard and prove-termination macro commands to make a final proof attempt if necessary, using the current-theory, and to make better use of the verbose argument.  Used those enhancements to improve APT's SIMPLIFY transformation by making that extra proof attempt when necessary (for guard verification and for proving termination); also improved output to report making additional additional proof attempts after the first when the :print option is :info or :all.
The key change is that the SIMPLIFY transformation can now succeed
where formerly it failed, because of the extra effort made in
verifying guards and termination (as discussed above).

proof-builder-macros.lisp:
Improved the prove-guard and prove-termination macro commands so that
if the first proof attempt fails, and so does the second attempt if
one is specified, then a third proof attempt is made using the
proof-builder's current theory.  Messages mentioning proof attempts
after the first are printed when the verbose argument is true.
The verbose argument now controls printing for all proof attempts,
not merely the first.