[ATC] Add initial support for function calls.
Now ATC accepts ACL2 functions that call other functions. Recursion is
disallowed for now. All the called functions must be among the explicit
targets. The targets must be ordered so that each target may only call preceding
targets. (Some of these restrictions may be relaxed at some point, but we also
strive to keep the code generator as simple as possible.)

The proof generation for now expands all function calls via symbolic execution.
This may be adequate for relatively few and small functions, but we plan to
refine this strategy. To make the symbolic execution work, it was necessary to
add some opener rules.