[ATC] Extend the C static semantics.
Add support for declarations in blocks.

This involve updating the static environment (specifically, the variable table)
from one block item to the next. A declaration updates the table, while a
statement does not.

It is allowed for a variable declaration in a block to hide one with the same
name in an outer block, but it is disallowed to hide one in the same block. This
is according to the rules of the C18 standard.