[ecurve] Formalize birational Montgomery-twistedEdwards point mapping.
For now this is only defined on the points for which the rational formulas
wwork, i.e. the denominators are not zero. There is a small finite number of
points on the curves for which the rational do not work: those have to mapped
specially; we will eventually add these mappings.

It also remains to prove that the resulting points are on the corresponding
curve (where 'corresponding' is defined by the mappings between curves --
vs. the mappings between points). Inversion theorems also remain to be proved.