[ecurve] Add a theorem about Montgomery curves.
This says that if A^2-4 is not a square, then (0, 0) is the only point on the
curve with y = 0.

This has significance, in particular, for the birational mappings with twisted
Edwards curves: points with y = 0 are exceptional (in the sense that they are
not amenable to the rational mappings because they make the denominators 0), and
this theorem tells us that, under the condition on A above, we only need to
worry about one point with y = 0.