[ecurve] Add Montgomery form of Jubjub.
We use the curve mapping to turn the twisted Edwards Jubjub into the Montgomery
Jubjub.

We prove a theorem saying that A^2 - 4 is not a square, where A is Jubjub's
Montgomery coefficient. This has applications to the theorem about the
exceptional points on a Montgomery curve for the birational mapping to twisted
Edwards.